[wasm][deoptimizer] Fix deopt loop for multi-instance cases
If there are two (or more) modules with disjoint call targets, whenever
tier-up is triggered, it is only ever going to use the feedback of one
wasm instance. To prevent deopt-loops, with this change the
TransitiveTypeFeedbackProcessor keeps the previous feedback "alive" by
also adding the old targets to the new feedback if present.

Bug: 42204618
Change-Id: I5bbc092f3d1ab964eb9ffe695a8521e4b3abc155