[parsing] Reuse scope infos held alive through eval
... by inserting the scope_info of the context in which we eval in
what used to be the "shared_function_infos" table. This CL renames
that table to the "infos" table. It would have been nice to add a
separate table, but mixing it into the same table has the advantage
that nested infos are easy to find.

As a drive-by change this change also drops `eval_scope_position` as a
way to distinguish eval. While it's technically fine for multiple
evals in the scope to share scripts, it actually breaks source
position info because the stack includes where eval was called by
attaching that position to the script. It's relatively unlikely that
we'll have multiple evals in the same scope with the same script
anyway...

Change-Id: I40fe2ed0c3fc9353e98846728561828f9803869d