[codegen] Remove RelocInfo::IsWasmReference
In most cases this predicate just mean "is there any reloc info", so
instead of checking for "wasm reference" explicitly, we now just check
if there is any mode set.

Note that the term "wasm reference" didn't carry a lot of meaning
anyway, it was just the conjunction of "wasm call" and "wasm canonical
signature ID".

R=nicohartmann@chromium.org

Bug: 42204513
Change-Id: I1ffeebb9169073ac54d547417cab533a07d5de4b