[wasm][sandbox] Harden WasmImportData against concurrent modification
While processing imports is specified to be an "atomic" operation
in the sense that no user code runs in the same Isolate at the same
time, a compromised worker could cause corruption anywhere in the
same sandbox cage at any time. That means we must not read the same
property repeatedly and trust that we get the same value; instead
we must read security-relevant values only once, and cache them
off-heap as long as they are needed.
This patch applies that idea to ProcessImportedFunction().
As a hardening measure, it also takes away a few helper functions
that are easy to misuse.

Fixed: 349529650
Bug: 336507783
Change-Id: Icba0d8159e1e5063c4c730120ebd6e397a4b67da