[maglev] loop speeling: Don't store optimistic loop type in Phi::type_
The problem with optimistically setting Phi::type_ is that at the end of
the loop we do not really know anymore what the type was at the
beginning. Thus we cannot reliably check if the type is consistent with
the loop end state.

Thus, this CL moves the optimistic phi type into the known node aspects
instead of the phi type. This essentially has the same effect but plays
nicely with our already existing consistency checks of known node aspect
states.

Fixed: 352402876
Change-Id: Id1cfe0a4f29f31f95032606c2ce44f7e8673db83