Reland "Support GetEnumeratedKeyedProperty bytecode in Maglev/Turbofan"
This is a reland of commit 19cf1a317f65f0b94eba7082400cfd2608e0df29

We should reset current_for_in_state.enum_cache_indices on resumable
loop headers. Because it is not stored in any local register, we will
lose this value in the loop body.

Original change's description:
> Support GetEnumeratedKeyedProperty bytecode in Maglev/Turbofan
>
> Bug: v8:14245
> Change-Id: I1e84623b45969b9efee02e5469b7a3ecf747972a
> Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5634288
> Reviewed-by: Leszek Swirski <leszeks@chromium.org>
> Commit-Queue: Hao A Xu <hao.a.xu@intel.com>
> Cr-Commit-Position: refs/heads/main@{#94956}

Bug: v8:14245
Change-Id: Iaf98f657056d925367b11bc53be249f7eadff31a