[MSan] Fix use-add-dtor error in GCTracer::ResetForTesting()
GCTracer::ResetForTesting() invokes the destructor and then uses
placement new to reinitialize GCTracer to its default state.
Unfortunately, the placement new call uses `heap_`, but `heap_` is no
longer valid to access after invoking the destructor.

The fix is to cache the value in a local and then pass the value of the
local to placement new.

Bug: 40222690
Change-Id: I64bce1a995fd648bda167f743a0f9f1ed8b34215