[heap] Fix tsan report about chunk metadata access
TSAN complains in crbug.com/353983482 that accessing the metadata of a
heap object in the JSHeapBroker is racy.
This seems like a false positive similar to how we have special TSAN
handling in `MemoryChunkMetadata::SynchronizedHeapLoad`. However, I
fixed it by doing a flag check first and only reading the metadata now
for supposedly read-only pages.
While I'm at it, inline the metadata table access to avoid an
unnecessary sbxcheck inside the metadata accessor.

Bug: 353983482
Change-Id: I9cfd7c2c826f972e00b0b7c75091dcafcbbb434b