[conversions] Clean up non-decimal prefix handling
StringToDouble is only ever called with all of, or none of,
ALLOW_HEX/BINARY/OCTAL, so unify these into ALLOW_NON_DECIMAL_PREFIX.
Also, it is never called with both ALLOW_NON_DECIMAL_PREFIX and
ALLOW_TRAILING_JUNK, so remove the bitflags and make this just a normal
enum. The reason this complexity used to exist has now been moved into a
separate parser (StringToIntHelper), so it's not needed in
StringToDouble anymore.

Additional small cleanups:

  * Use const Char* instead of an abstract Iterator, since that's all it
    ends up being
  * Make InternalStringToIntDouble responsible for marking an empty
    string as invalid -- this means StringToIntHelper has to instead
    handle the case where something looks like a legitimate radix
    prefix (i.e. leading zero), but turns out to actually be junk.
  * The prefix processing in StringToDouble preserves the character
    iterator for the case where there's no valid prefix, which
    simplifies the logic around processing sign/infinity vs leading
    zeros.

Change-Id: Iaf8333b76e2afb372e852b29e44b1432ec2e4d2c