[sandbox][leaptiering] Add dispatch table entries for builtins
On startup create a dispatch table entry for builtins needed in CSA
code. These cells are stored in IsolateData::builtins_dispatch_table.

The new `AllocateFunctionWithContext` CSA helper creates a JSFunction
for a builtin and correctly sets its dispatch entry.

The list of these builtins is in JS_BUILTIN_DISPATCH_HANDLE_ROOT_LIST
and still empty in this CL and will populated when needed.

Bug: 40931165, 42204201
Change-Id: Ia79ff4b03ee6d034a78a5219099147724f7c4476