Create Simplified Heap Iteration for Detached Elements
Design: https://docs.google.com/document/d/1b4jSkDB7T-3G16O_9OVN1XG-ET7UD5JpFnn0bxm3PTk/edit

Added a function to the HeapProfiler to enumerate the heap and filter for JSApiWrapperObjects. These objects are evaluated for detachedness and will be sent to Blink to find the top node of the detached tree.

This CL was accidentally uploaded as a cherry pick CL. I have been unable to convert this CL to a non-cherry-picked CL. Thus, a new CL has been created:
https://chromium-review.googlesource.com/c/v8/v8/+/5767438

Bug: 350519222
Change-Id: I5dd543b961a19fd3e3252eba2385b52127927be7