[etw] do not log MethodLoad events for builtins
These methods are embedded in the binary, and should be present
in the PDB file. Removing the trace events simplifies the process of
profiling this code on a Windows machine. These symbols, when loaded from the PDB, still appear in Windows Performance Analyzer stacktraces as "chrome.dll!Builtin_ObjectConstructor" (for example).

Change-Id: I0a834c99a6543e0364af49ed0cb0248b185e732c