[turboshaft][wasm][arm64] Fix crash in RecreateSchedule
The new optimization introduced in https://crrev.com/c/5736193 emits
a new operation that is neither supported by the RecreateSchedulePhase
nor by Turbofan (by design).

However, as the default still is to run with the Turbofan ISel even
with `--turboshaft-wasm`, our benchmarks hit the RecreateSchedule path
for Turboshaft benchmarks running on arm64.

This is fixed by skipping the optimization if the new Turboshaft ISel
is not enabled.

Fixed: 357946495
Change-Id: I60bd068c22c252bcae9af003eaea72351b2af286