[jumbo] Define DEBUG=1 in debug mode.
There are three common ways V8 code does preprocessor checks for debug
mode, in decreasing order of frequency:

1. #ifdef DEBUG
https://source.chromium.org/search?q=f:%5Ev8%2F%20%23ifdef%5C%20DEBUG
2. #if DEBUG
https://source.chromium.org/search?q=f:%5Ev8%2F%20%23if%5C%20DEBUG
3. #if defined(DEBUG)
https://source.chromium.org/search?q=f:%5Ev8%2F%20%23if%5C%20defined%5C(DEBUG%5C)

Currently BUILD.gn #defines DEBUG (to nothing), which will satisfy (1)
and (3), but not (2). #defining it to 1 satisfies all three.

Another option would be to mass-fix everyone to style (1), but it's
difficult to prevent backsliding, especially in more complicated
preprocessor expressions. This change is not only much easier, it seems
safer.

Bug: none
Change-Id: I2b391181ee09f0c8374f137a4b2ffb916d4a4c41