[NOT FOR COMMIT] Illustrate various kinds of tracing-related changes.
utils.h is the new utilities, while the other files are
randomly-selected files that use them. I picked files with larger deltas
in hopes of showing more "meaty" changes, but otherwise didn't look at
what was in these or if they cast the proposed changes in an especially
good or bad light.

Change-Id: Ie8e7efeab7885de5aee6923a35a6086d1ee3e0fc