[sandbox][leaptiering] Implement basic dispatch handle write barrier
For now, the write barrier only marks the table entry as alive if
necessary. In the future, it could either be extended to mark both the
table entry and the (Code) object referenced through it as alive, or we
could have two separate write barriers. The write barrier will be needed
in a follow-up CL when setting the dispatch handle on a FeedbackCell.

Bug: 40931165, 42204201
Change-Id: I421fc731a66ba60b85a43b0b4c0d6e7c6a0045c3