[sandbox][leaptiering] Allocate dispatch entries in JSFunctionBuilder
When allocating a new JSFunction object, we will try to reuse the
dispatch entry of the function's FeedbackCell. However, this isn't
possible for functions using the generic many_closures_cell. With this
CL, we will now allocate a new dispatch entry for those functions. That
requires ensuring that the SharedFunctionInfo object has the correct
parameter count, and therefore requires a few changes elsewhere.

Bug: 40931165, 42204201
Change-Id: I5892654b04c54e8822dfd9cef14ab7da5afce137