[maglev] Support skipping blocks in graph processor
...and use it to skip blocks which are not relevant for LICM.

The semantics of skipping are that any following graph processor is
skipped as well. We could also only skip if all processors want to skip.
But then we also need to support skipping individual processors within
the node, at which point it might not be beneficial anymore. Either way,
this can be figured out once we identify a second user of this api.

Bug: 350764035
Bug: v8:7700
Change-Id: Id65717bcbcaaf3d8f4053b5d7c930949ec606182