[maglev] Ensure object escapes when storing within try block
This is crucial because a deoptimization point inside the
catch handler could re-materialize objects differently,
depending on whether the throw occurred before or after
this store.

We could potentially relax this requirement by verifying
that no throwable nodes have been emitted since
the try-block started,  but for now, err on the side
of caution and always escape.

Fixed: 360182480
Change-Id: I19a61c6576053fdf847239edd73bdba802dd3ae6