[maglev] Add Branch to sub graph builder
It creates a version of SelectReduction that supports returning
multiple variables.

As previously, branches can abort, but it cannot fail.

Change manually crafted sub graph in
TryReduceArrayIteratorPrototypeNext to use the new
Branch method.

This fixes bug:360234501, since one of the branches
can now abort. This also allows one of the branches to
be completely elided when the condition is statically
known.

Fixed: 360234501
Change-Id: I786b7d300fb60ddca3f0a74fa71d9e96051099d9