[turboshaft][wasm] Fix handling InstanceCache in return_call_ref
on speculative inlining.
A similar issue exists for return_call_indirect which is going to be
addressed as part of https://crbug.com/360052650.
The speculative inlining for call_indirect also misses the
InstanceCache handling but only on polymorphic inlining for any
inlinee but the first (the slow path handles it already correctly).

Bug: 360044696
Change-Id: I63e19cb47bff255dd8f3739f1586d41c67d3bead