Merged: [wasm] Refactor central stack switches
Instead of storing the stack limit and stack pointer in a fixed slot of
the switching frame, store this information in the StackMemory. This
makes the logic independent of the frame type, which will make it easier
to extend to other code kinds, and in particular to wasm builtins.

Bug: 349640002,42202153
(cherry picked from commit 2419129afe69dba8ff5e6d55f1271d8c47c0e283)

Change-Id: I8bdac3fac56a25e905803bf094fc0813ba09164b