[top-level-await] Fix errored case in module graphs in TLA
In module graphs with TLA, it is possible that a module evaluation
causes the rejection of one of its async ancestors, setting the
ancestor's state to kErrored. When evaluation bubbles back up to that
evaluating that ancestor, it should skip that ancestor because it's
already errored. This case was incorrectly thought to be impossible by
construction and was not correctly handled.

Consider a graph with three (3) modules:

async_entry.mjs [[HasTLA]]
      | imports
      v
 sync_throw.mjs
      | imports
      v
async_leaf.mjs [[HasTLA]]

When async_leaf.mjs finishes async evaluating, it gathers its ancestors
[sync_throw.mjs, async_entry.mjs] and starts evaluating them.

First, it evaluates sync_throw.mjs, which doesn't have a TLA, and
synchronously throws. This causes it to reject async_entry.mjs, its
async ancestor, and set the state to kErrored.

Second, it evaluates async_entry.mjs, which is assumed to not be
kErrored, but is in fact already kErrored due to the previous step.

Also drive-by cleans up spec step comments.

Fixed: 357487162
Change-Id: Ieaeed44c0948fccefe8ec3f2651e7ec3e61add4c