[ic] Fix clone ic looking up cross context maps
We decided to not install cross-context maps into clone ICs, since the
ramifications were not fully understood.

However it's possible that such a cross context map is picked up
through the cache on the transition tree. Thus this needs to be exluded
as well to stay consistent.

Object.assign is not affected because it guards the fastcase with a
check of the source map against the object function from the current
native context.

Bug: 362017087
Change-Id: I36929f9ffa5b5d5885f7f1857b4c64cd3ddedea5