[ic] Introduce MegaTransition state for KeyedStoreIC
Initializing an object from empty object literal by a template object
can easily hit the megamorphic state in the process of adding a new
property. For example:

  function foo(config) {
    const props = {};
    for (var propName in config) {
      props[propName] = 0
    }
    return props;
  }

The megamorphic handler will try to lookup the property name in
descriptor array first and then in transition array. However, we can
avoid the descriptor array lookup if we know that this is a
transitioning store.

This CL firstly sets the inline cache of a keyed store to
MegaTransition state when there is only transition store. Then in
JIT code it will generate a call to Builtin_KeyedStoreIC_Transition
which will do transition array lookup first.

Change-Id: I291dc7077ac725f04f9b1afe25fd8086240adac8