[sandbox][leaptiering] Adjust dispatch handle write barrier
As dispatch handles are only processed during major GC, we can always
skip the write barrier during minor GC. This helps avoid scenarios where
a minor-GC marking barrier marks a table entry as alive but not the
contained Code object (because that's not a young-gen object). This in
turn could lead to a somewhat inconsistent state, which can then confuse
the code flushing logic (it would flush an unexpected code kind).

Bug: 40931165, 42204201
Change-Id: I844648fede59ad2208d61f58f52eaf66fb6c517d