Optimize JSON.stringify for simple large strings
JSON.stringify is slower than a user-land JavaScript function when dealing with strings that don’t contain characters requiring escaping. The difference becomes particularly significant, with a 20-30 times slowdown, for very long strings (e.g., >10,000,000 characters).

The cause of this was that the string was being scanned and copied one character at a time from the start to the end.

To address this, I made a change to copy the slice of the string from one escape position to the next, which significantly improved performance.

Bug: 40069803
Change-Id: I2b139d0903bf9f70e6b58e1a4a786d8511495ccb