[wasm] embed full pointers for inlining comparisons
Instead of loading the trusted instance => jump table start during
inlining checks, embed the full pointer of the jump table address.

This adds a new WASM_INDIRECT_CALL_TARGET relocation to make this
check relocatable.
Note that in a future CL, I'm planning to replace the embedded pointer
with an embedded handle to the WasmCodePointerTable.

Bug: 335082212
Change-Id: I66980144cc3ba95d54caea886de175702d0f80ed