[test] Fix integer overflow in test function
The test function sums up the elements in a TypedArray. The fuzzer
created an Int32Array such that adding the elements causes an integer
overflow, which is undefined behavior for int32_t. With this CL the test
function sums up the elements of the TypedArray as doubles, and then
casts back to the expected type to avoid the undefined behavior.

R=marja@chromium.org

Bug: 367779587
Change-Id: I9a7e3157afc3a84f4bb0fcb60c7fbe45b2cbcc81