[wasm] Introduce IndexType and use it for tables and memories
This introduces a new enum for the IndexType (i32 or i64). This replaces
multiple enums we were using before (WasmTableFlag,
TestingModuleMemoryType, MemoryType, TableType) and multiple boolean
fields (is_table64, is_memory64).

It's a lot of code churn, but pretty mechanical. This will allow us to
parameterize on the index type better.

R=evih@chromium.org

Bug: 358381633
Change-Id: I0dea6ade67df0445c19572a11cb7d476cf9575bb