[foozzie] Make compact error reporting the default
A while back, a compact error reporting mode was added to reduce the number of duplicates in case of bad clustering.

The --compact flag is passed to foozzie on Clusterfuzz for more than half a year now, so that we can deprecate the non-compact mode. There are no longer any open mixed cases.

This removes code for the previous non-compact output and simplifies the code paths as we don't longer need to differentiate.

No-Try: true
Bug: 369105515
Change-Id: Icc19be63d0aa08ccea5323e2b93ff3e48644b56a