[foozzie] Simplify configs, flags and output
Previously we used a configuration label and its associated list of flags. However, sometimes extra flags added to one side of the configuration make the difference and not the default flags of the comparison configuration.

Since the configuration label is only cosmetic, this changes the code, so that an experiment config is well-defined by the executable, the inferred architecture and the flags. The label of the configuration name is dropped to simplify the code.

This also simplifies the output and improves how flags are presented. Common flags (the majority) and flags from only one of the comparisons are now separated.

How failures are represented in Clusterfuzz is not changed as Clusterfuzz only parses the output header.

This enables regenerating test expectations using an environment variable (GENERATE=1). The variable was used to regenerate the files.

No-Try: true
Bug: 369105515
Change-Id: Ie301ef11a162db30c20323e6b91d1fdc8fe39a23