[wasm] Store canonicalized signature pointers on imports
And use them in the generic wasm-to-js wrapper instead of the on-heap
serialized signature, which can be manipulated.

Overall we now use canonicalized signatures a lot more. This often
allows to drop a {WasmModule*} which was only used to make sense of
module-specific types and signatures.

We now need to translate between canonical type IDs and canonical
signatures in several places. I left a few TODOs to clean this up in
follow-up CLs.

R=jkummerow@chromium.org

Bug: 354408144
Change-Id: I1e7b86845f1f03f26a1404004c847bf1d4ef2bff