[wasm-js][cleanup] Replace output parameters by return values
A little preparatory refactoring to avoid output parameters and instead
return std::optional values. This is the more modern and idiomatic way
to do this.

As a drive-by, "GetIntegerProperty" was renamed to "GetIndexValue", and
will be extended in a follow-up CL to accept BigInt values for i64 index
values.

Also two dead parameters (isolate and lower bound) are removed where
they are not needed or always a constant value.

None of this should cause any semantic change.

R=evih@chromium.org

Bug: 358381633
Change-Id: I10ead8079b15e14183ae9ca65e20ca64794c5676