[sandbox] Define kJavaScriptCallDispatchHandleRegister
We want to change the JavaScript ABI so that the caller passes the
dispatch handle (or directly the pointer to the dispatch entry) to the
callee. That is for example necessary for the callee to be able to
determine the correct argument count. This CL is a first step in that
direction: we now define a new kJavaScriptCallDispatchHandleRegister
which the caller must set to the dispatch handle of the called function.

Bug: 40931165
Change-Id: I6534143e9c4d5042ed84c869f1a0812e7cc5d765