[class] deserialize receiver for static initializer reparsing
If the function is a static initializer function, the function body spans
across the part of the class body that contains static initializers, which
may contain instance initializers in between, and those instance initializers
could access the receiver from the computed field keys. So always allocate
the receiver for static initializers to make it reparsable.

This partially reverts https://chromium-review.googlesource.com/c/v8/v8/+/5872642
because the new check was not enough to replace the check added by
https://chromium-review.googlesource.com/c/v8/v8/+/4573678 when the
nearest receiver scope of the class scope (where the this expression is
evaluated) after deserialization doesn't have receiver allocated for
its own.

Bug: 371237564
Change-Id: Idfafcabbcdb65a3c93f3ae4e438da395bf082db2