Implement capture state support for the v8 ETW provider
This change supports emitting source and method events to ETW on demand,
specifically when the isenabled flag of the ETW callback method is 2.

Relevant documentation is here:
https://learn.microsoft.com/en-us/windows/win32/api/evntprov/nc-evntprov-penablecallback

Since the events are produced from the isolate thread, the callback
now blocks during an isenabled==2 callback until the isolates have
written the events or a timeout expires.

Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5873795
Bug: 340621497