[api] Add new API to get the content and size of an ArrayBufferView
V8 provides APIs to access the data and length of an ArrayBufferView
individually. However, by having to do two API calls one either has to
create a HandleScope to store a reference to the ArrayBufferView after a
type check, or one has to do the type check twice. In a node.js micro
benchmark this causes a 100% regression compared to the new API that is
introduced with this CL, where data and size can be acquired with a
single API call.
Bug: 356158110
Change-Id: I0d58bdcff36992be824c9cdf2b2165a30ab335e6