[wasm] Don't tier up wrapper if signature depends on other instance
The wasm-to-js wrapper tierup currently does not handle signatures with
indexed types correctly if the WebAssembly instance from which the
JavaScript function is called is different than the WebAssembly instance
that imported the JavaScript function initially. With this CL the
wrapper tierup gets disabled in that case until tierup gets fixed
eventually.

R=clemensb@chromium.org

Bug: 371565065
Change-Id: I75ddeced30defea332cbeb1636d0f249e8ef3083