[tagged] Make FixedArray a HeapObjectLayout class
Move TaggedArrayBase and all its subclasses to be C++ object
definitions, inheriting from HeapObjectLayout. The TaggedArrayBase holds
both the length/capacity and the elements, with subclasses not expected
(or able) to hold extra fields -- the Shape can include an ExtraFields
mixin class for subclasses that need extra fields (like ArrayList),
which is injected using template magic between the length/capacity and
elements.

This requires some extensions to the C++ object layout capability:

  * Adding a TrustedObjectLayout for starting to port over
    TrustedObjects, since some TaggedArrays are trusted,
  * Fixing TaggedMember to work with non-main cages (also for trusted
    arrays),
  * More atomics support for TaggedMember,
  * Adding support for the ephemeron barrier to TaggedMember (in fact, I
    just made EphemeronHashTable call the write barrier directly instead
    of having a generic macro, it's a very special case and can be
    treated as such),
  * Making sure that HeapObjectLayout is non-movable/non-copyable, to
    catch issues where we were passing it by value due to `*this`,

Bug: 42202654
Change-Id: I23df96ef2815c4592438137b4d7a77d496e1fe52