[handles] Introduce distinct DirectHandle in non-css
Add a distinct DirectHandle class in non-direct-handle builds, which is
a thin wrapper around (Indirect)Handle. This allows users of
DirectHandle to immediately be able to tell whether the code will also
build with direct handles enabled. It also allows us to make
DirectHandle overloads non-conditional, since it is now a distinct type
from Handle and thus overloadable.

Same story for MaybeDirectHandle.

Bug: 42203211
Change-Id: Ieb6680290c237e822e1ea9eef3c981a3960cef54