[turboshaft] Maglev-to-ts: explicit Uint32-Int32 conversions
Maglev nodes all have a value_representation (which can be Uint32,
Int32, Float64...), which, among other things, is used to know how to
retag untagged nodes that flow into exception Phis. For instance,
0xffffffff could be either -1 or 4294967295 depending on whether it
is a Uint32 or a Int32.

So, during the translation from Maglev to TS, RecordRepresentation
records for each OpIndex used as input to an exception phi what is
the value_representation of the corresponding node in Maglev, so
that we know how to retag it.

However, this gets broken by some Turboshaft optimizations. For
instance, `x >>> 0` is optimized into `x` by Turboshaft, except that
its output is always unsigned, whereas its input `x` could be either
signed or unsigned. As a result, we can end up in a situation where
`x` in signed in maglev, and `x >>> 0` is unsigned, but Turboshaft
only has `x` in both cases, but it needs to somehow be both signed and
unsigned.
Similarly, because of GVN, 2 identical Word32 constants will be merged
in Turboshaft, even though one could be signed and the other one
unsigned.

So, this CL introduces explicit Signed/Unsigned type hints that should
be used when a node can return its input with a different sign (which
means that we could have a node with 2 different representations).


Note that https://crrev.com/c/5921411 already tried to solve this by
introducing an Identity node to hack around nodes having both Uint32
and Int32 representations, as it was assumed that uses of such node
would be rare. However, it turned out that there are a few cases
where nodes can have both Uint32 and Int32 representation, so this
hack became too inconvenient, hence the current CL that proposes a
cleaner approach.

Bug: chromium:42204525
Change-Id: I09a62e2f199dc80df9c780e4537cb371ca8bb9d9
Fixed: chromium:373025979