[float16array] Throw in Atomics methods
This CL fixes Atomics methods to throw when operating on Float16Arrays.

It does this by reordering the TypedArray ElementsKinds so that all
valid-for-atomic-access TA ElementsKinds are in a contiguous range that
comes before invalid-for-atomic-access ElementsKinds.

This CL also adds additional static and runtime safety checks.

Finally, drive-by removal of some dead methods.

Fixed: 372932101,373467913
Change-Id: I0bc72c403c225c44bcbf9bb59558c78039b2573f